1) Created folders for each class 2) changed .x and .y to sf::Vector2f in some places (to reduce lines of code) 3) For Gas - moved movement code from main.cpp to the class method movement();

+++ src/Barrier/Barrier.cpp
@@ -4,6 +4,6 @@
 
 #include "Barrier.h"
 
-Barrier::Barrier(int inx, int iny, int indx, int indy) // constructor
+Barrier::Barrier(float inx, float iny, float indx, float indy) // constructor
     : x(inx), y(iny), dx(indx), dy(indy), arr(nullptr) {}
